moved conveyor belt populator into environment startup
