+++ ariac_gazebo/config/trial_configuration/sample.yaml
@@ -32,7 +32,6 @@ parts:
         color: 'red'
         slots: [1, 5, 9]
         rotation: 'pi/6'
-        flipped: false
     bin3:
       - type: 'regulator'
         color: 'purple'
@@ -43,8 +42,8 @@ parts:
       - type: 'battery'
         color: 'blue'
         slots: [1, 3]
-        rotation: 0
-        flipped: true
+        rotation: 'pi/3'
+        flipped: false
       - type: 'battery'
         color: 'green'
         slots: [7, 9]
@@ -58,7 +57,7 @@ parts:
   conveyor_belt: # population params for conveyor belt
     active: true
     spawn_rate: 3.0 # seconds between spawn
-    order: 'random' # random or sequential
+    order: 'sequential' # random or sequential
     parts_to_spawn:
       - type: 'battery'
         color: 'red'
